// pro/src/apiClient/adage/models/CollectiveAdditionalFeeResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CollectiveAdditionalFeeType } from './CollectiveAdditionalFeeType';
export type CollectiveAdditionalFeeResponse = {
  amount: number;
  label?: string | null;
  type: CollectiveAdditionalFeeType;
};