// pro/src/apiClient/v1/models/GetCollectiveOfferLocationModelV2.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CollectiveLocationType } from './CollectiveLocationType';
import type { LocationResponseModelV2 } from './LocationResponseModelV2';
export type GetCollectiveOfferLocationModelV2 = {
  location: (LocationResponseModelV2 | null);
  locationComment: (string | null);
  locationType: CollectiveLocationType;
};